[TAN-1885] Move code into new AttendeesGenerator
