+++ back/app/controllers/web_api/v1/events_controller.rb
@@ -42,7 +42,7 @@ class WebApi::V1::EventsController < ApplicationController
     attendees = User.where(id: event.attendances.pluck(:attendee_id))
 
     I18n.with_locale(current_user&.locale) do
-      xlsx = XlsxService.new.generate_attendees_xlsx attendees, view_private_attributes: true
+      xlsx = XlsxExport::AttendeesGenerator.new.generate_attendees_xlsx attendees, view_private_attributes: true
       send_data xlsx, type: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet', filename: 'attendees.xlsx'
     end
   end
